+++ bindings/nodejs/src/client.rs
@@ -27,14 +27,7 @@ impl Finalize for ClientMethodHandlerWrapper {}
 impl ClientMethodHandler {
     pub fn new(channel: Channel, options: String) -> Result<Self> {
         let runtime = tokio::runtime::Runtime::new().expect("error initializing client");
-        let client = runtime
-            .block_on(
-                ClientBuilder::new()
-                    .from_json(&options)
-                    .expect("error initializing client")
-                    .finish(),
-            )
-            .expect("error initializing client");
+        let client = runtime.block_on(ClientBuilder::new().from_json(&options)?.finish())?;
 
         Ok(Self { channel, client })
     }