+++ sdk/src/wallet/message_interface/message_handler.rs
@@ -37,7 +37,7 @@ use crate::{
     wallet::{
         account::{
             operations::transaction::{
-                high_level::minting::mint_native_token::MintTokenTransactionDto, TransactionOptions,
+                high_level::minting::create_native_token::CreateNativeTokenTransactionDto, TransactionOptions,
             },
             types::{AccountIdentifier, TransactionDto},
             OutputDataDto,
@@ -671,14 +671,25 @@ impl WalletMessageHandler {
                     transactions.iter().map(TransactionDto::from).collect(),
                 ))
             }
-            AccountMethod::DecreaseNativeTokenSupply {
+            AccountMethod::CreateNativeToken { params, options } => {
+                convert_async_panics(|| async {
+                    let transaction = account
+                        .create_native_token(params, options.map(TransactionOptions::try_from_dto).transpose()?)
+                        .await?;
+                    Ok(Response::CreateNativeTokenTransaction(
+                        CreateNativeTokenTransactionDto::from(&transaction),
+                    ))
+                })
+                .await
+            }
+            AccountMethod::MeltNativeToken {
                 token_id,
                 melt_amount,
                 options,
             } => {
                 convert_async_panics(|| async {
                     let transaction = account
-                        .decrease_native_token_supply(
+                        .melt_native_token(
                             token_id,
                             U256::try_from(&melt_amount).map_err(|_| Error::InvalidField("melt_amount"))?,
                             options.map(TransactionOptions::try_from_dto).transpose()?,
@@ -688,33 +699,20 @@ impl WalletMessageHandler {
                 })
                 .await
             }
-            AccountMethod::IncreaseNativeTokenSupply {
+            AccountMethod::MintNativeToken {
                 token_id,
                 mint_amount,
                 options,
             } => {
                 convert_async_panics(|| async {
                     let transaction = account
-                        .increase_native_token_supply(
+                        .mint_native_token(
                             token_id,
                             U256::try_from(&mint_amount).map_err(|_| Error::InvalidField("mint_amount"))?,
                             options.map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
-                    Ok(Response::MintTokenTransaction(MintTokenTransactionDto::from(
-                        &transaction,
-                    )))
-                })
-                .await
-            }
-            AccountMethod::MintNativeToken { params, options } => {
-                convert_async_panics(|| async {
-                    let transaction = account
-                        .mint_native_token(params, options.map(TransactionOptions::try_from_dto).transpose()?)
-                        .await?;
-                    Ok(Response::MintTokenTransaction(MintTokenTransactionDto::from(
-                        &transaction,
-                    )))
+                    Ok(Response::SentTransaction(TransactionDto::from(&transaction)))
                 })
                 .await
             }